Rename hook file
